Make dbutils typecast use a valid type variable (#259)
## Changes
* We used an variable for collecting type information about dbutils.
This behaviour was patched in a recent release of pyright
(https://github.com/microsoft/pyright/commit/6169e0fb3cc966e530578b98fe24a395ed67e8e4).
Now we need to use an explicit type variable.
## Tests
- [x] tested manually on vscode with pylance `v2023.7.40` and type
checking mode set to `strict`
- [x] `make test` run locally
- [x] `make fmt` applied
- [x] relevant integration tests applied
