+++ databricks/sdk/runtime/__init__.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from typing import Union
+
 is_local_implementation = True
 
 # All objects that are injected into the Notebook's user namespace should also be made
@@ -33,21 +35,22 @@ try:
         _globals[var] = userNamespaceGlobals[var]
     is_local_implementation = False
 except ImportError:
-    from typing import Type, cast
+    from typing import cast
 
     # OSS implementation
     is_local_implementation = True
 
-    try:
-        from . import stub
-        from .stub import *
-        dbutils_type = Type[stub.dbutils]
-    except (ImportError, NameError):
-        from databricks.sdk.dbutils import RemoteDbUtils
+    from databricks.sdk.dbutils import RemoteDbUtils
 
+    from . import dbutils_stub
+
+    dbutils_type = Union[dbutils_stub.dbutils, RemoteDbUtils]
+
+    try:
+        from .stub import *
+    except (ImportError, NameError):
         # this assumes that all environment variables are set
         dbutils = RemoteDbUtils()
-        dbutils_type = RemoteDbUtils
 
     dbutils = cast(dbutils_type, dbutils)
